% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\docType{data}
\name{wikipedia}
\alias{wikipedia}
\title{Wikipedia Noun Articles}
\format{A data frame with 1,403,063 rows and 4 variables}
\usage{
data(wikipedia)
}
\description{
A dataset containing wikipedia articles searched for via a nouns data base.
}
\details{
\itemize{
  \item noun. The noun used to search for the Wikipedia article.
  \item text. The text from the article.
  \item element_id. The article number.
  \item sentence_id. The sentence within the \code{element_id}.
}
}
\keyword{datasets}

